Also defer #close_connection_after_writing
If only #send_data is deferred to the next tick, any call to
before the deferred writes could be executed. Now, this call is queued
just as #send_data is, to preserve the original ordering.

Because the SSL handshake can only occur when the server has sent an
info string to the client, initial writes may *not* be deferred. Writes
may be deferred when either SSL is not used, or when the SSL handshake
completes.